fixed #451,remove the code of support mac platform. 1. Remove the EventDispatcher. 2. Refactor CCLayer, CCMenu and CCParticleSystemPoint. 3. Modify the project file of win32, wophone and android.
