+++ cocos2dx/Android.mk
@@ -31,17 +31,16 @@ cocoa/CCSet.cpp \
 cocoa/CCZone.cpp \
 cocos2d.cpp \
 effects/CCGrabber.cpp \
-event_dispatcher/CCEventDispatcher.cpp \
-event_dispatcher/CCKeyboardEventDelegate.cpp \
-event_dispatcher/CCMouseEventDelegate.cpp \
 keypad_dispatcher/CCKeypadDelegate.cpp \
 keypad_dispatcher/CCKeypadDispatcher.cpp \
 label_nodes/CCLabelAtlas.cpp \
 label_nodes/CCLabelBMFont.cpp \
 label_nodes/CCLabelTTF.cpp \
+layers_scenes_transitions_nodes/CCLayer.cpp \
 layers_scenes_transitions_nodes/CCScene.cpp \
 layers_scenes_transitions_nodes/CCTransitionPageTurn.cpp \
 layers_scenes_transitions_nodes/CCTransitionRadial.cpp \
+menu_nodes/CCMenu.cpp \
 menu_nodes/CCMenuItem.cpp \
 misc_nodes/CCMotionStreak.cpp \
 misc_nodes/CCProgressTimer.cpp \
@@ -49,12 +48,11 @@ misc_nodes/CCRenderTexture.cpp \
 misc_nodes/CCRibbon.cpp \
 particle_nodes/CCParticleExamples.cpp \
 particle_nodes/CCParticleSystem.cpp \
+particle_nodes/CCParticleSystemPoint.cpp \
 particle_nodes/CCParticleSystemQuad.cpp \
 platform/CCDirector_mobile.cpp \
 platform/CCFileUtils.cpp \
 platform/CCGrid_mobile.cpp \
-platform/CCLayer_mobile.cpp \
-platform/CCMenu_mobile.cpp \
 platform/CCNode_mobile.cpp \
 platform/CCGL.cpp \
 platform/CCImage.cpp \
@@ -62,7 +60,6 @@ platform/CCStdC.cpp \
 platform/CCSAXParser.cpp \
 platform/CCThread.cpp \
 platform/CCCommon.cpp \
-platform/CCParticleSystemPoint_mobile.cpp \
 platform/CCTransition_mobile.cpp \
 platform/platform.cpp \
 platform/android/CCEGLView_android.cpp \
