+++ cocos2dx/actions/CCActionTiledGrid.cpp
@@ -1,5 +1,5 @@
 /****************************************************************************
-Copyright (c) 2010я╝Н2011 cocos2d-x.org
+Copyright (c) 2010гн2011 cocos2d-x.org
 Copyright (c) 2009       On-Core
  
 http://www.cocos2d-x.org
