+++ cocos2dx/menu_nodes/CCMenu.cpp
@@ -1,5 +1,6 @@
 /****************************************************************************
-Copyright (c) 2010 cocos2d-x.org
+Copyright (c) 2010-2011 cocos2d-x.org
+Copyright (c) 2008-2010 Ricardo Quesada
 
 http://www.cocos2d-x.org
 
@@ -23,14 +24,14 @@ THE SOFTWARE.
 ****************************************************************************/
 #include "CCMenu.h"
 #include "CCDirector.h"
-#include "CCXApplication.h"
-#include "CGPointExtension.h"
+#include "CCApplication.h"
+#include "CCPointExtension.h"
 #include "CCTouchDispatcher.h"
 #include "CCTouch.h"
-#include "platform/CCXMath.h"
+#include "CCStdC.h"
 
 #include <vector>
-#include <float.h>
+
 using namespace std;
 
 namespace cocos2d{
@@ -55,7 +56,7 @@ namespace cocos2d{
 			return pRet;
 		}
 		va_end(args);
-		CCX_SAFE_DELETE(pRet)
+		CC_SAFE_DELETE(pRet)
 		return NULL;
 	}
 
@@ -63,10 +64,10 @@ namespace cocos2d{
 	{
 		if (CCLayer::init())
 		{
-			this->m_bIsMouseEnabled = true;
+			this->m_bIsTouchEnabled = true;
 
 			// menu in the center of the screen
-			CGSize s = CCDirector::sharedDirector()->getWinSize();
+			CCSize s = CCDirector::sharedDirector()->getWinSize();
 
 			this->m_bIsRelativeAnchorPoint = false;
 			m_tAnchorPoint = ccp(0.5f, 0.5f);
@@ -74,6 +75,17 @@ namespace cocos2d{
 
 			// XXX: in v0.7, winSize should return the visible size
 			// XXX: so the bar calculation should be done there
+			CCRect r;
+            CCApplication::sharedApplication().statusBarFrame(&r);
+			ccDeviceOrientation orientation = CCDirector::sharedDirector()->getDeviceOrientation();
+			if (orientation == CCDeviceOrientationLandscapeLeft || orientation == CCDeviceOrientationLandscapeRight)
+			{
+				s.height -= r.size.width;
+			}
+			else
+			{
+				s.height -= r.size.height;
+			}
 			setPosition(ccp(s.width/2, s.height/2));
 
 			int z=0;
@@ -102,16 +114,16 @@ namespace cocos2d{
 	/*
 	* override add:
 	*/
-	CCNode * CCMenu::addChild(CCNode * child, int zOrder)
+	void CCMenu::addChild(CCNode * child, int zOrder)
 	{
-		return CCLayer::addChild(child, zOrder);
+		CCLayer::addChild(child, zOrder);
 	}
 
-	CCNode * CCMenu::addChild(CCNode * child, int zOrder, int tag)
+	void CCMenu::addChild(CCNode * child, int zOrder, int tag)
 	{
 		// we can not use RTTI, so we do not known the type of object
-		/*NSAssert( dynamic_cast<CCMenuItem*>(child) != NULL, L"Menu only supports MenuItem objects as children");*/
-		return CCLayer::addChild(child, zOrder, tag);
+		/*CCAssert( dynamic_cast<CCMenuItem*>(child) != NULL, L"Menu only supports MenuItem objects as children");*/
+		CCLayer::addChild(child, zOrder, tag);
 	}
 
     void CCMenu::onExit()
@@ -126,131 +138,66 @@ namespace cocos2d{
         CCLayer::onExit();
     }
 
-    void CCMenu::registerWithTouchDispatcher()
-    {
-        CCLOG("cocos2d: CCMenu: unsupported");
-    }
-
-    CCMenuItem* CCMenu::itemForTouch(CCTouch * touch)
-    {
-        CCLOG("cocos2d: CCMenu: unsupported");
-        return NULL;
-    }
-
-    bool CCMenu::ccTouchBegan(CCTouch* touch, UIEvent* event)
-    {
-        CCLOG("cocos2d: CCMenu: unsupported");
-        return false;
-    }
-
-    void CCMenu::ccTouchEnded(CCTouch* touch, UIEvent* event)
-    {
-        CCLOG("cocos2d: CCMenu: unsupported");
-    }
-
-    void CCMenu::ccTouchCancelled(CCTouch *touch, UIEvent* event)
-    {
-        CCLOG("cocos2d: CCMenu: unsupported");
-    }
-
-    void CCMenu::ccTouchMoved(CCTouch* touch, UIEvent* event)
-    {
-        CCLOG("cocos2d: CCMenu: unsupported");
-    }
-
 	//Menu - Events
-    int CCMenu::mouseDelegatePriority()
-    {
-        return -1;/** @todo upto-0.99.5  use NSIntegerMin+1 instead*/
-    }
+	void CCMenu::registerWithTouchDispatcher()
+	{
+		CCTouchDispatcher::sharedDispatcher()->addTargetedDelegate(this, kCCMenuTouchPriority, true);
+	}
 
-    CCMenuItem* CCMenu::itemForMouseEvent(NSEvent * pEvent)
-    {
-        CGPoint touchLocation = CCDirector::sharedDirector()->convertEventToGL(pEvent);
+	bool CCMenu::ccTouchBegan(CCTouch* touch, CCEvent* event)
+	{
+		if (m_eState != kCCMenuStateWaiting || ! m_bIsVisible)
+		{
+			return false;
+		}
+		m_pSelectedItem = this->itemForTouch(touch);
+		if (m_pSelectedItem)
+		{
+			m_eState = kCCMenuStateTrackingTouch;
+			m_pSelectedItem->selected();
+			return true;
+		}
+		return false;
+	}
 
-        if (m_pChildren && m_pChildren->count() > 0)
-        {
-            NSMutableArray<CCNode*>::NSMutableArrayIterator it;
-            for (it = m_pChildren->begin(); it != m_pChildren->end(); ++it)
-            {
-                if (! *it)
-                {
-                    break;
-                }
+	void CCMenu::ccTouchEnded(CCTouch *touch, CCEvent* event)
+	{
+		CCAssert(m_eState == kCCMenuStateTrackingTouch, "[Menu ccTouchEnded] -- invalid state");
+		if (m_pSelectedItem)
+		{
+			m_pSelectedItem->unselected();
+			m_pSelectedItem->activate();
+		}
+		m_eState = kCCMenuStateWaiting;
+	}
 
-                if ((*it)->getIsVisible() && ((CCMenuItem*)(*it))->getIsEnabled())
-                {
-                    CGPoint local = (*it)->convertToNodeSpace(touchLocation);
+	void CCMenu::ccTouchCancelled(CCTouch *touch, CCEvent* event)
+	{
+		CCAssert(m_eState == kCCMenuStateTrackingTouch, "[Menu ccTouchCancelled] -- invalid state");
+		if (m_pSelectedItem)
+		{
+			m_pSelectedItem->unselected();
+		}
+		m_eState = kCCMenuStateWaiting;
+	}
 
-                    CGRect r = ((CCMenuItem*)(*it))->rect();
-                    r.origin = CGPointZero;
-
-                    if (CGRect::CGRectContainsPoint(r, local))
-                    {
-                        return (CCMenuItem*)(*it);
-                    }
-                }
-            }
-        }
-
-        return NULL;
-    }
-
-    bool CCMenu::ccMouseUp(NSEvent * pEvent)
-    {
-        if (m_pSelectedItem)
-        {
-            m_pSelectedItem->unselected();
-            m_pSelectedItem->activate();
-
-            m_eState = kCCMenuStateWaiting;
-            return true;
-        }
-        
-        return false;
-    }
-
-    bool CCMenu::ccMouseDown(NSEvent * pEvent)
-    {
-        if (! getIsVisible())
-        {
-            return false;
-        }
-        
-        m_pSelectedItem = itemForMouseEvent(pEvent);
-        m_pSelectedItem->selected();
-
-        if (m_pSelectedItem)
-        {
-            m_eState = kCCMenuStateTrackingTouch;
-            return true;
-        }
-        
-        return false;
-    }
-
-    bool CCMenu::ccMouseDragged(NSEvent * pEvent)
-    {
-        CCMenuItem* currentItem = itemForMouseEvent(pEvent);
-
-        if (currentItem && currentItem != m_pSelectedItem)
-        {
-            if (m_pSelectedItem)
-            {
-                m_pSelectedItem->unselected();
-            }
-
-            m_pSelectedItem = currentItem;
-            m_pSelectedItem->selected();
-        }
-        
-        if (currentItem && m_eState == kCCMenuStateTrackingTouch)
-        {
-            return true;
-        }
-        
-        return false;
-    }
+	void CCMenu::ccTouchMoved(CCTouch* touch, CCEvent* event)
+	{
+		CCAssert(m_eState == kCCMenuStateTrackingTouch, "[Menu ccTouchMoved] -- invalid state");
+		CCMenuItem *currentItem = this->itemForTouch(touch);
+		if (currentItem != m_pSelectedItem) 
+		{
+			if (m_pSelectedItem)
+			{
+				m_pSelectedItem->unselected();
+			}
+			m_pSelectedItem = currentItem;
+			if (m_pSelectedItem)
+			{
+				m_pSelectedItem->selected();
+			}
+		}
+	}
 
 	void CCMenu::destroy(void)
 	{
@@ -273,7 +220,7 @@ namespace cocos2d{
 		float height = -padding;
 		if (m_pChildren && m_pChildren->count() > 0)
 		{
-			NSMutableArray<CCNode*>::NSMutableArrayIterator it;
+			CCMutableArray<CCNode*>::CCMutableArrayIterator it;
 			for (it = m_pChildren->begin(); it != m_pChildren->end(); ++it)
 			{
                 if (!(*it))
@@ -288,7 +235,7 @@ namespace cocos2d{
 		float y = height / 2.0f;
 		if (m_pChildren && m_pChildren->count() > 0)
 		{
-			NSMutableArray<CCNode*>::NSMutableArrayIterator it;
+			CCMutableArray<CCNode*>::CCMutableArrayIterator it;
 			for (it = m_pChildren->begin(); it != m_pChildren->end(); ++it)
 			{
                 if (!(*it))
@@ -313,7 +260,7 @@ namespace cocos2d{
 		float width = -padding;
 		if (m_pChildren && m_pChildren->count() > 0)
 		{
-            NSMutableArray<CCNode*>::NSMutableArrayIterator it;
+            CCMutableArray<CCNode*>::CCMutableArrayIterator it;
 			for (it = m_pChildren->begin(); it != m_pChildren->end(); ++it)
 			{
                 if (!(*it))
@@ -328,7 +275,7 @@ namespace cocos2d{
 		float x = -width / 2.0f;
 		if (m_pChildren && m_pChildren->count() > 0)
 		{
-            NSMutableArray<CCNode*>::NSMutableArrayIterator it;
+            CCMutableArray<CCNode*>::CCMutableArrayIterator it;
 			for (it = m_pChildren->begin(); it != m_pChildren->end(); ++it)
 			{
                 if (!(*it))
@@ -369,7 +316,7 @@ namespace cocos2d{
 
 		if (m_pChildren && m_pChildren->count() > 0)
 		{
-			NSMutableArray<CCNode*>::NSMutableArrayIterator it;
+			CCMutableArray<CCNode*>::CCMutableArrayIterator it;
 			for (it = m_pChildren->begin(); it != m_pChildren->end(); ++it)
 			{
 				// if it has no value, break
@@ -385,7 +332,7 @@ namespace cocos2d{
 				assert(rowColumns);
 
 				float tmp = (*it)->getContentSize().height;
-				rowHeight = (unsigned int)((rowHeight >= tmp || CCXMath::isnanCocos2d(tmp)) ? rowHeight : tmp);
+				rowHeight = (unsigned int)((rowHeight >= tmp || isnan(tmp)) ? rowHeight : tmp);
 
 				++columnsOccupied;
 				if (columnsOccupied >= rowColumns)
@@ -402,7 +349,7 @@ namespace cocos2d{
 		// check if too many rows/columns for available menu items
 		assert(! columnsOccupied);
 
-		CGSize winSize = CCDirector::sharedDirector()->getWinSize();
+		CCSize winSize = CCDirector::sharedDirector()->getWinSize();
 
 		row = 0;
 		rowHeight = 0;
@@ -413,7 +360,7 @@ namespace cocos2d{
 
 		if (m_pChildren && m_pChildren->count() > 0)
 		{
-			NSMutableArray<CCNode*>::NSMutableArrayIterator it;
+			CCMutableArray<CCNode*>::CCMutableArrayIterator it;
 			for (it = m_pChildren->begin(); it != m_pChildren->end(); ++it)
 			{
 				if (! *it)
@@ -429,7 +376,7 @@ namespace cocos2d{
 				}
 
 				float tmp = (*it)->getContentSize().height;
-				rowHeight = (unsigned int)((rowHeight >= tmp || CCXMath::isnanCocos2d(tmp)) ? rowHeight : tmp);
+				rowHeight = (unsigned int)((rowHeight >= tmp || isnan(tmp)) ? rowHeight : tmp);
 
 				(*it)->setPosition(ccp(x - winSize.width / 2,
 					                   y - (*it)->getContentSize().height / 2));
@@ -481,7 +428,7 @@ namespace cocos2d{
 
 		if (m_pChildren && m_pChildren->count() > 0)
 		{
-			NSMutableArray<CCNode*>::NSMutableArrayIterator it;
+			CCMutableArray<CCNode*>::CCMutableArrayIterator it;
 			for (it = m_pChildren->begin(); it != m_pChildren->end(); ++it)
 			{
 				if (! *it)
@@ -498,7 +445,7 @@ namespace cocos2d{
 
 				// columnWidth = fmaxf(columnWidth, [item contentSize].width);
 				float tmp = (*it)->getContentSize().width;
-				columnWidth = (unsigned int)((columnWidth >= tmp || CCXMath::isnanCocos2d(tmp)) ? columnWidth : tmp);
+				columnWidth = (unsigned int)((columnWidth >= tmp || isnan(tmp)) ? columnWidth : tmp);
 
 				columnHeight += (int)((*it)->getContentSize().height + 5);
 				++rowsOccupied;
@@ -520,7 +467,7 @@ namespace cocos2d{
 		// check if too many rows/columns for available menu items.
 		assert(! rowsOccupied);
 
-		CGSize winSize = CCDirector::sharedDirector()->getWinSize();
+		CCSize winSize = CCDirector::sharedDirector()->getWinSize();
 
 		column = 0;
 		columnWidth = 0;
@@ -530,7 +477,7 @@ namespace cocos2d{
 
         if (m_pChildren && m_pChildren->count() > 0)
 		{
-			NSMutableArray<CCNode*>::NSMutableArrayIterator it;
+			CCMutableArray<CCNode*>::CCMutableArrayIterator it;
 			for (it = m_pChildren->begin(); it != m_pChildren->end(); ++it)
 			{
 				if (! *it)
@@ -546,7 +493,7 @@ namespace cocos2d{
 
 				// columnWidth = fmaxf(columnWidth, [item contentSize].width);
 				float tmp = (*it)->getContentSize().width;
-				columnWidth = (unsigned int)((columnWidth >= tmp || CCXMath::isnanCocos2d(tmp)) ? columnWidth : tmp);
+				columnWidth = (unsigned int)((columnWidth >= tmp || isnan(tmp)) ? columnWidth : tmp);
 
 				(*it)->setPosition(ccp(x + columnWidths[column] / 2,
 					                   y - winSize.height / 2));
@@ -575,7 +522,7 @@ namespace cocos2d{
 
 		if (m_pChildren && m_pChildren->count() > 0)
 		{
-			NSMutableArray<CCNode*>::NSMutableArrayIterator it;
+			CCMutableArray<CCNode*>::CCMutableArrayIterator it;
 			for (it = m_pChildren->begin(); it != m_pChildren->end(); ++it)
 			{
 				if (! *it)
@@ -603,7 +550,7 @@ namespace cocos2d{
 
 		if (m_pChildren && m_pChildren->count() > 0)
 		{
-			NSMutableArray<CCNode*>::NSMutableArrayIterator it;
+			CCMutableArray<CCNode*>::CCMutableArrayIterator it;
 			for (it = m_pChildren->begin(); it != m_pChildren->end(); ++it)
 			{
 				if (! *it)
@@ -624,4 +571,40 @@ namespace cocos2d{
 	{
 		return m_tColor;
 	}
+
+	CCMenuItem* CCMenu::itemForTouch(cocos2d::CCTouch *touch)
+	{
+		CCPoint touchLocation = touch->locationInView(touch->view());
+		touchLocation = CCDirector::sharedDirector()->convertToGL(touchLocation);
+
+        if (m_pChildren && m_pChildren->count() > 0)
+		{
+			CCMutableArray<CCNode*>::CCMutableArrayIterator it;
+			for (it = m_pChildren->begin(); it != m_pChildren->end(); ++it)
+			{
+				if (! *it)
+				{
+					break;
+				}
+
+				// ignore invisible and disabled items: issue #779, #866
+				if ((*it)->getIsVisible() && ((CCMenuItem*)(*it))->getIsEnabled())
+				{
+					CCPoint local = (*it)->convertToNodeSpace(touchLocation);
+
+					CCRect r = ((CCMenuItem*)(*it))->rect();
+					r.origin = CCPointZero;
+
+					if (CCRect::CCRectContainsPoint(r, local))
+					{
+						return (CCMenuItem*)(*it);
+					}
+				}
+			}
+			
+		}
+
+		return NULL;
+	}
+
 }
