+++ cocos2dx/particle_nodes/CCParticleSystemPoint.cpp
