+++ cocos2dx/platform/CCArchOptimalParticleSystem.h
@@ -30,7 +30,6 @@
 // ARMv7, Mac or Simulator use "Quad" particle
 #if defined(__ARM_NEON__) ||  \
 	(CC_TARGET_PLATFORM == CC_PLATFORM_AIRPLAY) || \
-	(CC_TARGET_PLATFORM == CC_PLATFORM_MAC)		|| \
 	(CC_TARGET_PLATFORM == CC_PLATFORM_IOS)		|| \
 	(CC_TARGET_PLATFORM == CC_PLATFORM_WIN32)	|| \
 	(CC_TARGET_PLATFORM == CC_PLATFORM_WOPHONE) || \
